clean up controllers file
