+++ controllers/all-artwork-routes.js
@@ -29,16 +29,6 @@ router.get('/', async (req, res) => {
                         }
                     ]
                 },
-                attributes: [
-                    'id',
-                    'title',
-                    'medium',
-                    'created_at',
-                    'date',
-                    'style',
-                    'location',
-                    'image_url'
-                ],
                 include: [
                     {
                         model: Museum,
@@ -52,16 +42,6 @@ router.get('/', async (req, res) => {
             })
         } else {
             artwork = await Artwork.findAll({
-                attributes: [
-                    'id',
-                    'title',
-                    'medium',
-                    'created_at',
-                    'date',
-                    'style',
-                    'location',
-                    'image_url'
-                ],
                 include: [
                     {
                         model: Museum,
@@ -78,7 +58,7 @@ router.get('/', async (req, res) => {
 
         //pass a single post object into the homepage
         const posts = artwork.map(post => post.get({ plain: true }))
-        res.render('all-artwork', {posts});
+        res.render('all-artwork', {posts, loggedIn: req.session.loggedIn});
 
     } catch (error) {
         console.log('Failed to get posts');
@@ -87,51 +67,5 @@ router.get('/', async (req, res) => {
         
     }    
 });
-
-router.get('/single-artwork/:id', (req, res) => {
-    Artwork.findOne({
-        where: {
-            id: req.params.id
-        },attributes: [
-            'id',
-            'title',
-            'medium',
-            'created_at',
-            'date',
-            'style',
-            'location',
-            'image_url'
-        ],
-        include: [
-            {
-                model: Museum,
-                attributes: ['username']
-            },
-            {
-                model: Artist,
-                attributes: ['artist_name']
-            }
-        ]
-    })
-    .then(dbPostData => {
-        if (!dbPostData) {
-            res.status(404).json({ message: 'No post found with this id' });
-            return;
-        }
-
-        // serialize the data
-        const post = dbPostData.get({ plain: true });
-
-        // pass data to template
-        //can we change to single-artwork
-        res.render('single-artwork', {
-            post
-        });
-    })
-    .catch(err => {
-        console.log('failed to get post');
-        res.status(500).json(err);
-    });
-})
 
 module.exports = router;
