+++ controllers/dashboard-routes.js
@@ -8,16 +8,6 @@ router.get('/', (req, res) => {
       where: {
         museum_id: req.session.museum_id
       },
-      attributes: [
-        'id',
-        'title',
-        'medium',
-        'created_at',
-        'date',
-        'style',
-        'location',
-        'image_url'
-      ],
       include: [
         {
           model: Museum,
@@ -40,53 +30,5 @@ router.get('/', (req, res) => {
         res.status(500).json(err);
     });
 });
-
-//this route gives the user the ability to edit a post
-router.get('/edit/:id', (req, res) => {
-    Artwork.findOne({
-      where: {
-        id: req.params.id
-      },
-      attributes: [
-        'id',
-        'title',
-        'medium',
-        'created_at',
-        'date',
-        'style',
-        'location',
-        'image_url'
-      ],
-      include: [
-        {
-          model: Museum,
-          attributes: ['username']
-        },
-        {
-          model: Artist,
-          attributes: ['artist_name']
-        }
-      ]
-    })
-      .then(dbPostData => {
-        if (!dbPostData) {
-          res.status(404).json({ message: 'No post found with this id' });
-          return;
-        }
   
-        // serialize the data
-        const post = dbPostData.get({ plain: true });
-  
-        // pass data to template
-        res.render('edit-post', {
-          post,
-          loggedIn: true
-        });
-      })
-      .catch(err => {
-        console.log('failed to edit posts');
-        res.status(500).json(err);
-      });
-  })
-  
-  module.exports = router;
+module.exports = router;
