+++ views/homepage.handlebars
@@ -13,3 +13,6 @@
             <button class="btn" id="search-artist">Search</button>
         </form>
 </section>
+
+
+<script src="/javascript/search-bar.js"></script>
