+++ views/layouts/dashboard.handlebars
@@ -112,5 +112,6 @@
 <script src="/javascript/add-artwork.js"></script>
 <script src="/javascript/search-bar.js"></script>
 <script src="/javascript/single-artwork.js"></script>
+<script src="/javascript/logout.js"></script>
 
 </html>
