+++ views/partials/artwork.handlebars
@@ -1,7 +1,7 @@
 <div class="row art-list">
     <div class="col-sm-3">
         <div class="card-deck" style="max-width: 18rem;">
-            <div class="card id={{id}}">
+            <div class="card" id="{{id}}">
                 <img src="{{image_url}}" class="card-img-top card-image" alt="artwork images">
                 <div class="card-body">
                     <h5 class="card-title">{{title}}</h5>
